Add a cooldownInterval options
When monitoring a script, it's handy to limit the maximum number of
restart. But it's not the same if the script fails 10 times in 10
seconds and 10 times in 10 weeks. On the first case, there is probably
an hard error and it's better to stop soon trying to restart this
script. But on the second case, a max of 10 is not enough. Maybe it's
just a crontab running once a week that makes it fail.

The cooldownInterval is here to make max time sensitive. Forever-monitor
keeps a counter with the number of failures of the script. When this
counter reaches the max, it stops trying to restart it. With
cooldownInterval, this counter is halfed at a regular interval defined
by the number of seconds of cooldownInterval.

For example, with cooldownInterval=300, if the counter is at 4 failures
and in the next 5 minutes, the script won't fail, it will go to 2. And
if the next 5 following minutes, it still run quiet, it will go to 1.
And 5 more minutes, and will be reset to 0.