Winston log integration and log rotation features
This commit uses Winston logging for forever logFile, outFile and
errFile. It's possible to use the log rotation features of winston (see
tailable, maxsize, maxFiles and zippedArchive options).
See https://github.com/foreverjs/forever/issues/128
Note that log rotation for the logFile will not work if the spawner of
this process (forever) has provided a file descriptor for stdout and
stderr, those values must be setted to 'ignore' and will be subject of a
PR to the forever repository